+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * 2.0.next in progress
   * Allow `:compile-ns` to accept regex strings to match namespaces (as well as symbols). PR #67 @wandersoncferreira (bartuka)
   * Fix #66 by switching from "jar" processing to "zip" processing and using `.entries` instead of calling `.getNextEntry` (the latter checks CRCs, the former doesn't apparently).
+  * Address #65 by ignoring `.DS_Store` files.
   * Fix #64 by adding a `:jvm-opts` exec argument for passing JVM options to the AOT compilation subprocess.
   * Fix #63 by adding a `:compile-fn` exec argument for passing in a custom `compile` function.